move group tests method to utils
